+++ scripts/api/sports-game-odds/sports.js
@@ -1,0 +1,25 @@
+/**
+ * @fileoverview Query sports data from Sports Game Odds.
+ */
+import 'dotenv/config';
+import axios from "axios";
+
+/**
+ * Query sports data.
+ */
+export async function getSports() {
+
+  const options = {
+    method: "GET",
+    url: "https://api.sportsgameodds.com/v2/leagues/",
+    // eslint-disable-next-line no-undef
+    params: { apiKey: process.env.SPORTS_GAME_ODDS },
+  };
+
+  try {
+    const { data } = await axios.request(options);
+    return data;
+  } catch (error) {
+    console.error(error);
+  }
+}
