+++ scripts/api/the-odds-api/events.js
@@ -20,8 +20,8 @@ export function getEvents(sportKey) {
     })
     .then((response) => {
       // Check your usage
-      console.log("Remaining requests", response.headers["x-requests-remaining"]);
-      console.log("Used requests", response.headers["x-requests-used"]);
+      console.log("Remaining Requests =>", response.headers["x-requests-remaining"]);
+      console.log("Used Requests =>", response.headers["x-requests-used"]);
       return response;
     })
     .catch((error) => {
