+++ spec/sports-game-odds/sports.spec.js
@@ -3,10 +3,11 @@
  * @fileoverview Test the query to retrieve sports data.
  */
 import 'dotenv/config';
-import * as Sports from "../../scripts/extract/sports-game-odds/sports.js" ; 
+import * as Sports from "../../scripts/api/sports-game-odds/sports.js" ; 
 
 
 describe("Sports Game Odds - Retrieve sport data", function () {
+
   it("Retrieve sport data", async function () {
 
     // Query Sports
@@ -14,8 +15,9 @@ describe("Sports Game Odds - Retrieve sport data", function () {
 
       // Check if the sports are not undefined
       expect(response.data).not.toBeUndefined();
-      
+
     });
 
   });
+
 });
