+++ spec/the-odds-api/sports.spec.js
@@ -3,7 +3,7 @@
  * @fileoverview Test the query to retrieve sports data.
  */
 import 'dotenv/config';
-import * as Sports from "../../scripts/extract/the-odds-api/sports.js" ; 
+import * as Sports from "../../scripts/api/the-odds-api/sports.js" ; 
 
 
 describe("The Odds API - Retrieve sports data", function () {
